alter weight from integer to decimal

+++ account/models.py
@@ -8,7 +8,8 @@ class Profile(models.Model):
     profile_pic = models.CharField(null=True, blank=True, max_length=255)
     # location = models.CharField(max_length=30, blank=True)
     birth_date = models.DateField(null=True, blank=True)
-    weight = models.IntegerField(null=True, blank=True)
+    # weight = models.IntegerField(null=True, blank=True)
+    weight = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     height = models.IntegerField(null=True,blank=True)
 
     def __str__(self):
